flask_app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
requests==2.32.3
js2py==0.74
lxml==5.2.2
python-dotenv==1.0.1
selenium==4.23.1
webdriver-manager==4.0.2
